refactor: change api name to events
