+++ events/templates/events/list.html
