+++ events/urls.py
@@ -1,6 +1,6 @@
 from django.urls import include, path, re_path
 from rest_framework import routers
-from api import views
+from events import views
 
 
 router = routers.DefaultRouter()
